cleanup and split up utils
