+++ .gitignore
@@ -6,3 +6,4 @@ build
 cmake-*
 prebuilds
 report.*.json
+dist
